requirements.txt
himl
pyyaml~=5.4.1
tqdm~=4.62.1
tensorflow
pandas~=1.3.2
scikit-learn~=0.24.2
opencv-python~=4.5.3.56
pillow
matplotlib~=3.4.2
seaborn~=0.11.2
numpy~=1.20.3
albumentations